Eliminate a couple of unnecessary classes
